app/requirements.txt
django==2.2
coverage>=4.4
django_coverage_plugin
psycopg2-binary
gunicorn
tblib
djangorestframework
drf-nested-routers
django-cors-headers
coreapi
Markdown
Pygments